Delete src/main/java/ca/mcmaster/se2aa4/island/teamXXX/DynamicPosition.java
Deleted the DynamicPosition file. It's better to keep the logic of changing direction and flying in the drone class since it is the drone who is in control of its position and direction and the drone is responsible for that. If we want to segregate this from the actual drones in the future potentially because of more features to be added, we can create a DisplacementSubsystem class which sole responsibility is to manage the movement of the drone. But based on the current business specifications and the interface provided to us, I believe that is not necessary since we don't have a low enough interaction with the hardware.